Large single commit because I had Git issues.  Solves the code challenge, as well as updating the file structure for the repo and adding READMEs.
