+++ code-challenges/C#/BinarySearch/BinarySearch.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.28307.489
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BinarySearch", "BinarySearch\BinarySearch.csproj", "{16758859-93C1-4913-867D-307E75F6F1BA}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "XUnitTestBinarySearch", "XUnitTestBinarySearch\XUnitTestBinarySearch.csproj", "{C446A978-EAE3-4F16-BE65-491F4F456175}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{16758859-93C1-4913-867D-307E75F6F1BA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{16758859-93C1-4913-867D-307E75F6F1BA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{16758859-93C1-4913-867D-307E75F6F1BA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{16758859-93C1-4913-867D-307E75F6F1BA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C446A978-EAE3-4F16-BE65-491F4F456175}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C446A978-EAE3-4F16-BE65-491F4F456175}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C446A978-EAE3-4F16-BE65-491F4F456175}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C446A978-EAE3-4F16-BE65-491F4F456175}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {A1DDBC0E-37EA-4FB8-AADC-F0E7679F0E47}
+	EndGlobalSection
+EndGlobal
