+++ code-challenges/C#/BinarySearch/README.md
@@ -1,0 +1,11 @@
+# Binary Search
+A short code challenge done with a whiteboard and implemented in C#.  This task was done in partnership with Meggan Triplett
+
+## Challenge
+Write a function called BinarySearch which takes in 2 parameters: a sorted array and the search key. Without utilizing any of the built-in methods available to your language, return the index of the array’s element that is equal to the search key, or -1 if the element does not exist.
+
+## Approach & Efficiency
+For this challenge we took a very straightforward approach.  We drew out a couple of sample arrays and illustrated what we would check as we itterated accross each one, and what the outcome would be if we did or did not find the search key.  We then wrote up a quick O(N) method to complete the challenge.
+
+## Solution
+<!-- Embedded whiteboard image -->
